requirements.txt
discord
requests
python-dotenv
pillow
argon2-cffi
numpy
torch